refactor: org list in utils
